EL - GR - Update
Ghange log:
--Updated the lang file, based the latest template
--Changed the Medium size to 8

Note:
For making the new sub-title to fit, i changed the Medium size, to 8, but this maked also the results to be somehow small. I am suggesting, for if the same issue exits with other translation, to add a diffrent size changer, for the sub-title (Results based on currently known requirements! Older Hardware Supported if Data Wiped!) .